プロフィールページエラー

+++ app/routes/web.php
@@ -29,15 +29,17 @@ Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
-
+//ログインしているユーザー限定のルート
 Route::group(['middleware' => 'auth'], function() {
 
     // ユーザ関連
     Route::resource('users', 'App\Http\Controllers\UsersController', ['only' => ['index', 'show', 'edit', 'update']]);
 
     //フォローする
-    Route::post('users/{user}/follow', 'App\Http\Controllers\UsersController@follow')->name('follow');
+    Route::post('users/{user:user_id}/follow', 'App\Http\Controllers\UsersController@follow')->name('follow');
     //フォロー解除する
-    Route::delete('users/{user}/unfollow', 'App\Http\Controllers\UsersController@unfollow')->name('unfollow');
+    Route::delete('users/{user:user_id}/unfollow', 'App\Http\Controllers\UsersController@unfollow')->name('unfollow');
+
+
 
 });
